requirements.txt
# POP3 to Gmail Importer - Python Dependencies
# Python 3.9 or higher required

# Environment variable management
python-dotenv==1.0.0

# Gmail API libraries (v3.0)
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0
